chore(graphql): refactor auth backends
