+++ {{cookiecutter.github_repository}}/{{cookiecutter.main_module}}/graphql/backends.py
@@ -1,0 +1,22 @@
+from django.contrib.auth import get_user_model
+from {{cookiecutter.main_module}}.users.auth import tokens, backends
+
+UserModel = get_user_model()
+
+
+class GraphJWTAuthenticationBackend(backends.BaseJWTAuthenticationMixin):
+    def authenticate(self, request, **kwargs):
+        auth_res = super().authenticate(request)
+        try:
+            user, _ = auth_res
+        except TypeError:
+            user = auth_res
+        if user:
+            return user
+        return None
+
+    def get_user(self, user_id):
+        try:
+            return UserModel._default_manager.get(pk=user_id)
+        except UserModel.DoesNotExist:
+            return None
